Create a temporary table that will hold the learning content
In a temporary table so that it will be automatically dropped at the
end of the transaction. We need to have the list of skills at hand
to massively insert data. This commit does not include the values to
insert (because, you know, we don't have to disclose the learning
content to the public) ;) .